pyproject.toml
[tool.poetry]
name = "tic-tac-toe"
version = "0.1.0"
description = "Tic-tac-toe game bot"
authors = ["Andrei Donskoi <donskoi.ae@gmail.com>"]
packages = [
    { include = "tic-tac-toe", from = "." },
]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
numpy = "^1.26.2"
python-telegram-bot = "^20.7"
python-dotenv = "^1.0.0"

[tool.poetry.dev-dependencies]
pytest = "^7.4.3"
black = "^23.12.1"
pylint = "^3.0.3"
flake8 = "^7.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"